Fix: Detection Updater Bug
Fixed a bug where the detection grid records weren't updating as they were supposed to.

+++ src/utility_functions/quadratic_solver.rs
@@ -2,9 +2,9 @@
 
 use super::integer_math::square_root;
 
-pub fn solve_quadratic (a_x_1000: i64, b_x_1000:i64, c_x_1000:i64) -> (i64, i64) {
+pub fn solve_quadratic (a_x_1000: i128, b_x_1000:i128, c_x_1000:i128) -> (i64, i64) {
 
-    return ((-b_x_1000 + square_root(b_x_1000 * b_x_1000 - 4 * a_x_1000 * c_x_1000 )) / (2 * a_x_1000 ) , (-b_x_1000 - square_root(b_x_1000 * b_x_1000 - 4 * a_x_1000 * c_x_1000 )) / (2 * a_x_1000 ))
+    return (((-b_x_1000 + square_root(b_x_1000 * b_x_1000 - 4 * a_x_1000 * c_x_1000 )) / (2 * a_x_1000 )) as i64 , ((-b_x_1000 - square_root(b_x_1000 * b_x_1000 - 4 * a_x_1000 * c_x_1000 )) / (2 * a_x_1000 )) as i64)
 
     
 }
